+++ CHANGELOG.md
@@ -9,7 +9,9 @@ Added:
 
 Changed:
 * GUI Item icons rotated
-* Fixed ticket machine name typo updated render
+* Fixed ticket machine name typo and updated texture
+* Texture UV update: Fluoresecnt Lamp in ground state and extending state
+* Texture UV update: Platform Gates Upper part
 
 0.9
 ---